+++ src/main/java/com/example/preproj/service/EmailService.java
@@ -1,6 +1,5 @@
 package com.example.preproj.service;
 
-import com.example.preproj.model.Request;
 import com.example.preproj.model.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -10,7 +9,6 @@ import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -42,41 +40,25 @@ public class EmailService {
         emailSender.send(simpleMailMessage);
     }
 
-    public List<User> allRequest() {
-        List<Request> list = requestService.reqUsers();
-        List<User> userList = new ArrayList<>();
-        for (Request request : list) {
-            userList.add(userService.show(request.getUserId()));
-        }
-        return userList;
-    }
-
-    public boolean show(int id) {
-        return requestService.show(id) != null;
-    }
-
-    public void request(int id) {
+    public void mailAdmin(int id) {
         User user = userService.show(id);
         String message = user.getName() + " " + user.getLastName() + " wants to become an admin";
         List<User> allUsers = userService.allUsers();
-        requestService.add(id);
-            for (User u : allUsers) {
-                if (u.getRoles().toString().contains("ROLE_ADMIN")) {
-                    emailService.sendSimpleEmail(u.getEmail(), "Hello", message);
-                }
+        for (User u : allUsers) {
+            if (u.getRoles().toString().contains("ROLE_ADMIN")) {
+                emailService.sendSimpleEmail(u.getEmail(), "Hello", message);
             }
+        }
     }
 
     public void acceptedApplication(int id) {
         String message = "Your application is accepted!";
         mail(id, message);
-        requestService.update(id);
     }
 
     public void rejectedApplication(int id) {
         String message = "Your application has been rejected";
         mail(id, message);
-        requestService.delete(id);
     }
 
     public void mail(int id, String message) {
